Add parallel variants of LeMAC for VAES-enabled CPUs
Only for benchmarking purposes.

Parallel variants run multiple LeMAC instances simultaneously:

- 2 instances for AVX2+VAES
- 4 instances for AVX512

Each instance runs with a distinct initial state.

Context separation is achieved by setting the top half of the
counter (used for initializing states and keys) to
`<number of lanes> || <lane number> || 0*`
similar to the approach used in the parallel variants of AEGIS.

With a parallelism degree of 1, the initialization function is
identical to that of regular LeMAC.

Finalization is performed by absorbing the tags from all lanes
into the first lane, and computing the final tag from this
consolidated first lane using the standard LeMAC function.

These constructions are not intended for practical deployment
but are useful for measuring the performance impact of leveraging
VAES instructions.